refactor(model): Lotto 클래스 model 디렉토리로 이동
